feat(variant page): #SKFP-109 filter refactoring
